refactor: Update stepMia function to include log_name parameter
This commit updates the stepMia function in __init__.py to include a new parameter called log_name. This parameter is used to specify the name of the log file and should be obtained using the logger.get_log_filename() function. The path_log parameter has also been renamed to path_url_log for clarity. These changes improve the clarity and flexibility of the stepMia function.